+++ src/App.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { BrowserRouter } from "react-router-dom";
 import "./App.css";
 
-import RoutesComponent from "./components/routes-component";
+import RoutesComponent from "./components/routes/routes-component";
 import { ThemeProvider } from "@mui/material/styles";
 import CssBaseline from "@mui/material/CssBaseline";
 import { darkTheme, lightTheme } from "./constants/themes";
